requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
python-dotenv
requests
pandas_datareader